[LinearMediaPlayer] Video should fast seek in response to scrubbing https://bugs.webkit.org/show_bug.cgi?id=275277 rdar://129418990
Reviewed by Jer Noble.

LinearMediaKit only dispatches seek(to:from:metadata:) once at the end of a scrub, which means that
the video content is not updated in real-time during the scrubbing event. However,
seekThumbnail(to:) *is* dispatched in real-time while scrubbing. While the intent of this method is
to update the contents of our Playable object's thumbnailLayerPublisher, since we don't currently
support thumbnail layers we can use this method to fast seek the video content instead. The result
of this change is that scrubbing in LinearMediaPlayer works like it does in modern-media-controls,
where video content updates to keyframes in real-time as the user scrubs.

* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
(-[WKLinearMediaPlayerDelegate linearMediaPlayer:seekThumbnailToTime:]):

Canonical link: https://commits.webkit.org/279840@main