[UnifiedPDF] Some compositing tests are consistently failing. https://bugs.webkit.org/show_bug.cgi?id=270092 rdar://123634997
Reviewed by NOBODY (OOPS!).

When the compile time flag for UnifiedPDF was enabled on EWS we started
to consistently fail certain tests. This required a little more work
than a simple rebaselining so this patch brings them back in line with
the current status of UnifiedPDF. I also took the liberty of moving
these tests from the compositing/plugins/pdf directory to the pdf/unifiedpdf
directory.

A small code change was needed such that we return a contents size of
{ 0, 0 } if we do not have an installed PDF document. The reason for
this is because we could perform a document layout without an actual
PDF document installed and still set the document size using the document
and page margins in PDFDocumentLayout. The document margin would also
be used as the max row width and along with the available width to compute a document
fitting scale. Later on when we  would try to get the contents
size we would end up computing a size that required  scrollbars
even though we did not have a PDF installed. This caused us to incorrectly
create and report that we had a scrollable area.

* LayoutTests/compositing/plugins/pdf/pdf-in-embed-expected.txt: Removed.
* LayoutTests/compositing/plugins/pdf/pdf-in-styled-embed-expected.txt: Removed.
* LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree-dynamic.html:
* LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree-expected.txt: Removed.

* LayoutTests/pdf/unifiedpdf/compositing/pdf-in-embed-expected.txt: Added.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-in-embed.html: Renamed from LayoutTests/compositing/plugins/pdf/pdf-in-embed.html.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-in-styled-embed-expected.txt: Added.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-in-styled-embed.html: Renamed from LayoutTests/compositing/plugins/pdf/pdf-in-styled-embed.html.

These tests needed to be slightly rewritten so that they could take
advantage of the registerPDFTest API exposed by internals so that
that test was only run after the PDF document was installed in the
plugin. After that it was just a matter of rebaselining the expectations.

* LayoutTests/pdf/unifiedpdf/compositing/pdf-scrolling-tree-dynamic-expected.txt: Renamed from LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree-dynamic-expected.txt.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-scrolling-tree-dynamic.html: Renamed from LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree-dynamic.html.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-scrolling-tree-expected.txt: Added.
* LayoutTests/pdf/unifiedpdf/compositing/pdf-scrolling-tree.html: Renamed from LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree.html.

There were two main changes to these tests. The first was to rewrite
pdf-scrolling-tree in the same way that pdf-scrolling-tree-dynamic was
written to check for a certain number of scrolling areas and
then check the scrolling state tree. Otherwise we would not get
the correct information when checking the scrolling state tree.

We also needed to change the height of the embed since it completely
fit inside the plugin and as a result was not scrollable.

* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::contentsSize const):