API URL change request can be cancelled by JS navigation https://bugs.webkit.org/show_bug.cgi?id=274745
Reviewed by NOBODY (OOPS!).

The problem can be reproduced with very quick (in each 10ms) fragment navigation
(JS call of window.location.href). In such case, it is not possible to navigate to
other site via "API URL change request".

To solve the problem, this change cancels the fragment navigation if there is an ongoing
"API URL change request".

* Source/WebCore/loader/FrameLoader.cpp:
(WebCore::FrameLoader::loadURL):