REGRESSION (275396@main): [iOS] Safari sometimes crashes under -[WKContentView textStylingAtPosition:inDirection:] https://bugs.webkit.org/show_bug.cgi?id=275339 rdar://127571335
Reviewed by Aditya Keerthi.

After the change in 275396@main, it's possible for `-textStylingAtPosition:inDirection:` to crash
when attempting to insert a `nil` value into the resulting dictionary. This is because
`+[UIFont fontWithDescriptor:size:]` can apparently return a `nil` `UIFont`, but our code just
assumes that it always returns a non-null value.

Harden against this possibility by null-checking the font before trying to add it to the result, and
also augment a test to exercise this mitigation by forcing `+fontWithDescriptor:size:` to return
`nil` in the middle of the test, and verify that we handle that case gracefully by returning the
original font (without the new symbolic traits).

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textStylingAtPosition:inDirection:]):
* Tools/TestWebKitAPI/Tests/ios/AutocorrectionTestsIOS.mm:
(returnNil):
(TEST(AutocorrectionTests, FontAtCaretWhenUsingUICTFontTextStyle)):

Canonical link: https://commits.webkit.org/279907@main