Remove `ipc/dirty-region-overflow.html` https://bugs.webkit.org/show_bug.cgi?id=275333 rdar://129023377
Reviewed by Pascoe.

This test no longer reveals an issue without the fix in 278840@main because of IPC serialization changes.
Keeping this test working would mean repeatedly updating the test to account for updates in class layout
and IPC messages, so we should just remove it.

* LayoutTests/ipc/dirty-region-overflow-expected.txt: Removed.
* LayoutTests/ipc/dirty-region-overflow.html: Removed.
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/279916@main