[WPE][GTK] Un-deprecate webkit_back_forward_list_item_get_title() https://bugs.webkit.org/show_bug.cgi?id=275298
Reviewed by Adrian Perez de Castro.

We changed our minds and are bringing back the ability to set page
titles in the back/forward list. Apparently this functionality worked on
macOS.

It does *not* work for WPE/GTK. The page title is always equal to the
URL. I have not investigated to find out why. But we should still not
discourage applications from using this API if it might work in the
future.

* Source/WebKit/UIProcess/API/glib/WebKitBackForwardListItem.cpp:
(webkit_back_forward_list_item_get_title):
* Source/WebKit/UIProcess/API/glib/WebKitBackForwardListItem.h.in:
* Tools/MiniBrowser/gtk/BrowserWindow.c:
(browserWindowCreateBackForwardMenu):
* Tools/TestWebKitAPI/Tests/WebKitGLib/TestBackForwardList.cpp:
(testBackForwardListNavigation):
(testWebKitWebViewSessionState):

Canonical link: https://commits.webkit.org/279918@main