[WebGPU] error in WGSL::prepare with attached WGSL source https://bugs.webkit.org/show_bug.cgi?id=275406 rdar://129691678
Reviewed by Mike Wyrzykowski and Dan Glastonbury.

The generated `_ArrayLength` variable's `@binding` was colliding with user-defined
variables. The issue was that the `@binding` was calculated based in the last
variable in binding group, but the group wasn't sorted, so the last variable might
not contain the biggest binding for the group.

* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::collectGlobals):
* Source/WebGPU/WGSL/tests/valid/array-length-unordered.wgsl: Added.

Canonical link: https://commits.webkit.org/279987@main