[visionOS wk2] pointerevents/ios/inputmode-change-update-keyboard-after-pointerup.html is a constant text failure https://bugs.webkit.org/show_bug.cgi?id=275432 rdar://129786718
Reviewed by Wenson Hsieh.

The test in question was previously expecting the keyboard rect to have
a positive height on form control activation. This does not make sense
on visionOS since keyboard rects should have zero size in said platform.

This patch updates the test to have different expectations across
visionOS and iOS. To do so, we introduce the `keyboardAppearsOverContent`
property on the test runner. The property is named to reflect how it is
consulted in this test.

We also add a new baseline for visionOS.

* LayoutTests/platform/visionos/pointerevents/ios/inputmode-change-update-keyboard-after-pointerup-expected.txt: Added.
* LayoutTests/pointerevents/ios/inputmode-change-update-keyboard-after-pointerup.html:
* Tools/WebKitTestRunner/InjectedBundle/Bindings/TestRunner.idl:
* Tools/WebKitTestRunner/InjectedBundle/TestRunner.h:
(WTR::TestRunner::keyboardAppearsOverContent const):

Canonical link: https://commits.webkit.org/279990@main