REGRESSION(279662@main): [ iOS 18 ] 6 TestWebKitAPI.UnifiedTextReplacement tests are constant failure https://bugs.webkit.org/show_bug.cgi?id=275466 rdar://129560001
Reviewed by Megan Gardner.

When receiving replacement text from Writing Tools in the Smart Reply case, a DocumentFragment is
created from the attributed string, which is then used in a `ReplaceSelectionCommand` with `MatchStyle::Yes`.

In this case, if there are internal style attributes involved, then an assertion failure will be reached.

Fix by removing this assertion, since it is perfectly valid to have internal attributes in this case,
and the case is handled fine already anyways.

* Source/WebCore/css/MutableStyleProperties.cpp:
(WebCore::MutableStyleProperties::setProperty):

Canonical link: https://commits.webkit.org/280000@main