REGRESSION (UI-side compositing) Disabled scrollbars don't show in some cases https://bugs.webkit.org/show_bug.cgi?id=261295 rdar://115137778
Reviewed by Simon Fraser.

For the case where we need to show disabled scrollbars on a scroller without a layer we need
to have the scrollbars controller of type ScrollbarsControllerMac, so that we can create
a temporary NSScrollerImp in the web process to display the disabled scrollbars (similar to
what we do for svg foreign object, etc.). However, the previous fix checked usesAsyncScrolling
to determine this, but it is better to check usesCompositedScrolling (as for this case
usesAsyncScrolling returns true even when we don't have a layer). For the normal case we will
properly update the scrollbars controller type to RemoteScrollbarsController for when the layer
is made, and for this case where the layer isn't made, we will properly use ScrollbarsControllerMac.

* LayoutTests/fast/scrolling/mac/scrollbars/scrollbars-controller-type-expected.txt:
* LayoutTests/fast/scrolling/mac/scrollbars/scrollbars-controller-type.html:
* Source/WebCore/platform/ScrollableArea.h:
* Source/WebCore/testing/Internals.cpp:
(WebCore:: const):
(WebCore::Internals::scrollbarsControllerTypeForNode const):
* Source/WebCore/testing/Internals.h:
* Source/WebKit/WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::ensureScrollbarsController const):

Canonical link: https://commits.webkit.org/280026@main