Upgrade GCC requirement to 11.2.0 https://bugs.webkit.org/show_bug.cgi?id=275507 rdar://129866801
Reviewed by Yusuke Suzuki.

This patch bumps the minimum required GCC version to 11.2.0.

Based on the WebKitGTK dependency policy [1], support for Debian Bullseye
ended on 6/10/24. This means that we can now cater to the default GCC
version for Ubuntu 22.04 LTS as our minimum, which happens to be 11.2.0.
We plan to support this version till 4/25/25.

This version bump allows us to start building the WebKit against the C++23
standard, which we are tracking in bug webkit.org/b/263122.

[1]: https://docs.webkit.org/Ports/WebKitGTK%20and%20WPE%20WebKit/DependenciesPolicy.html

* Source/cmake/WebKitCommon.cmake:

Canonical link: https://commits.webkit.org/280033@main