Sec-Fetch-Destination is invalid for service worker intercepted loads https://bugs.webkit.org/show_bug.cgi?id=275474 rdar://129832604
Reviewed by Alex Christensen.

We are updating destination when empty based on the resource type.
We are then using destination to compute sec-fetch-dest which has a specific handling of empty.
To cover the case of mismatch (which happens only for service worker intercepted fetches), we are storing whether the load is a fetch.
In which case, we are using empty string as a the value of sec-fetch-dest.
.
* LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/navigation-headers.https-expected.txt:
* Source/WebCore/loader/cache/CachedResourceLoader.cpp:
(WebCore::updateRequestFetchMetadataHeaders):

Canonical link: https://commits.webkit.org/280070@main