AX: Notify accessibility that WebProcess is suspended on both Mac and iOS https://bugs.webkit.org/show_bug.cgi?id=275457 rdar://129796427
Reviewed by Andres Gonzalez and Brady Eidson.

https://bugs.webkit.org/show_bug.cgi?id=228936 (Notify accessibility
that WebProcess is suspended when in process cache) added an
accessibility notification that a process is suspended, but only for
iOS. We need the same notification on macOS also.

The Mac implementation of the notification is already there, and it's
already used in other places in the code where a process is
suspended. We were just missing the case where the process gets added
to the process cache.

The bug this causes is that assistive technology can try to query
attributes of an element that's in the suspended process. The query
hangs until it times out. This was happening sometimes when following
a link to a different domain with VoiceOver. The notification enables
accessibility clients to keep track of which processes are suspended
and cause any queries to those processes to fail fast.

* Source/WebKit/WebProcess/WebProcess.cpp:
(WebKit::WebProcess::prepareToSuspend):
(WebKit::WebProcess::processDidResume):

Canonical link: https://commits.webkit.org/280077@main