[results.webkit.org] ⌘F should not override system find-in-page if the search panel is collapsed https://bugs.webkit.org/show_bug.cgi?id=275531
Reviewed by BJ Burg.

Avoid preventing default and attempting to focus the search field when the user presses ⌘F, if the
search field is disabled (and therefore cannot be edited in the first place). The existing logic
makes it difficult to invoke the browser's built-in find-in-page functionality at all in
results.webkit.org, which can be useful in some cases (e.g. when searching for specific trains, OS
build versions, etc.).

* Tools/Scripts/libraries/resultsdbpy/resultsdbpy/view/static/js/drawer.js:

Also drive-by-fix a typo in a nearby local variable.

Canonical link: https://commits.webkit.org/280092@main