Make WeakPtr to Nodes thread safe again https://bugs.webkit.org/show_bug.cgi?id=275575
Reviewed by Andres Gonzalez.

Make WeakPtr to Nodes thread safe again. This shows as performance neutral on the benchmarks
we track. The idea was to make the WeakPtrImpl ref count thread safe so that we could use
WeakPtr<Node> in more places, where performance is critical. However, this didn't work out
and we plan to use CheckedPtr<Node> for those cases instead.

Note that being able to pass WeakPtr<Node> to another thread is important for accessibility
code as well.

* Source/WebCore/dom/EventTarget.h:

Canonical link: https://commits.webkit.org/280097@main